Fixes #38
You'd imagine that the `opcache_get_configuration` method would return all of the settings, but no.  So now we get the opcache config and the relevant ini settings for the extension and bring in any missing ones.  Not using the ini settings for everything as it returns the values as strings and not the various scalar types as befitting the value.