Fixes #37
Having the `file_cache_only` ini setting as '1' stops `opcache_get_statistics` returning a bunch of information.  It doesn't (understandably) return any of the memory info, but also doesn't seem to return a list of cached files, start/restart times, or any of the other statistics.  So this needs compensating for in the interface which leaves basically viewing the ini settings and links to the available functions.