Reset link now respect live updating
If the live updates are turned on then the reset tab will send an ajax request rather than reload the page.  The reset indicator has the same treatment as the live update (goes green by default and pulses) until the next live update where it refreshes the content and removes the 'working' indicator from the reset.